Make app more singleton friendly
Took 19 minutes
