+++ crates/dream-fs/src/js_fs.rs
@@ -1,5 +1,7 @@
 use std::path::PathBuf;
 
+use indexed_db_futures::prelude::*;
+use indexed_db_futures::IdbDatabase;
 use wasm_bindgen::prelude::*;
 
 use crate::fs::{FileKind, ReadDir};
@@ -14,28 +16,28 @@ extern "C" {
 
 #[allow(dead_code)]
 pub async fn read_binary_from_web_storage(file_path: &str) -> Vec<u8> {
-    // {
-    //     // Open my_db v1
-    //     let mut db_req: OpenDbRequest = IdbDatabase::open_u32("my_db", 1).expect("Err");
-    //     db_req.set_on_upgrade_needed(Some(|evt: &IdbVersionChangeEvent| -> Result<(), JsValue> {
-    //         // Check if the object store exists; create it if it doesn't
-    //         if let None = evt.db().object_store_names().find(|n| n == "my_store") {
-    //             evt.db().create_object_store("my_store").expect("Err");
-    //         }
-    //         Ok(())
-    //     }));
-    //
-    //     let db: IdbDatabase = db_req.into_future().await.expect("Err");
-    //
-    //     // Insert/overwrite a record
-    //     let tx: IdbTransaction = db
-    //         .transaction_on_one_with_mode("my_store", IdbTransactionMode::Readwrite)
-    //         .expect("Err");
-    //     let store: IdbObjectStore = tx.object_store("my_store").expect("Err");
-    //     let value_to_put: JsValue = JsValue::from("bar");
-    //     store.put_key_val_owned("foo", &value_to_put).expect("Err");
-    //     tx.await.into_result().expect("Err");
-    // }
+    {
+        // Open my_db v1
+        let mut db_req: OpenDbRequest = IdbDatabase::open_u32("my_db", 1).expect("Err");
+        db_req.set_on_upgrade_needed(Some(|evt: &IdbVersionChangeEvent| -> Result<(), JsValue> {
+            // Check if the object store exists; create it if it doesn't
+            if let None = evt.db().object_store_names().find(|n| n == "my_store") {
+                evt.db().create_object_store("my_store").expect("Err");
+            }
+            Ok(())
+        }));
+
+        let db: IdbDatabase = db_req.into_future().await.expect("Err");
+
+        // Insert/overwrite a record
+        let tx: IdbTransaction = db
+            .transaction_on_one_with_mode("my_store", IdbTransactionMode::Readwrite)
+            .expect("Err");
+        let store: IdbObjectStore = tx.object_store("my_store").expect("Err");
+        let value_to_put: JsValue = JsValue::from("bar");
+        store.put_key_val_owned("foo", &value_to_put).expect("Err");
+        tx.await.into_result().expect("Err");
+    }
 
     js_sys::Uint8Array::from(readBinary(file_path)).to_vec()
 }
