+++ crates/dream-window/src/lib.rs
@@ -1,1 +1,3 @@
+#![feature(async_closure)]
+
 pub mod window;
